infra/video-generate-service/requirements.txt
celery==5.4.0
minio==7.1.12
psycopg2-binary==2.9.9
python-dotenv==1.0.1